refactor: move the forward simulation helpers from utils to simulations
