+++ src/elicito/__init__.py
@@ -553,11 +553,13 @@ class Elicit:
                 variable.assign(weights[f"weight_{j}"].sel(replication=i).values)
 
             tf.random.set_seed(run_seed)
-            (elicits, prior_sim, model_sim, target_quants) = utils.simulate_and_elicit(
-                prior_model=prior_model,
-                model=self.model,
-                targets=self.targets,
-                seed=run_seed,
+            (elicits, prior_sim, model_sim, target_quants) = (
+                simulations.simulate_and_elicit(
+                    prior_model=prior_model,
+                    model=self.model,
+                    targets=self.targets,
+                    seed=run_seed,
+                )
             )
             simulated.append(
                 dict(
